refactor ctld grpc server
